Dmcrorie

+++ OneBook/app/src/main/java/com/example/onebook/Borrower.java
@@ -2,7 +2,7 @@ package com.example.onebook;
 
 public class Borrower extends User {
 
-    Borrower(String username, String password, String email, long phone, int userID){
+    Borrower(String username, String password, String email, long phone, long userID){
             super(username, password, email, phone, userID);
     }
 
@@ -10,7 +10,7 @@ public class Borrower extends User {
         super(username, password, email);
     }
 
-    public boolean requestBook(Book book, int userID) {   
+    public boolean requestBook(Book book, long userID) {
 
         if (book.getStatus().equals("available") && book.getOwner().getUserID() == userID) {
             return true;
@@ -19,7 +19,7 @@ public class Borrower extends User {
             return false;
     }
 
-    public boolean returnBook(Book book, int userID) {
+    public boolean returnBook(Book book, long userID) {
 
         //code to be added
         return true;
